Fix admin link and move some component files.
