+++ src/components/utils/IsAdmin.js
@@ -1,5 +1,5 @@
 export const IsAdmin = async (userId) => {
-    const isAdmin = await fetch(`http://localhost:8000/user/${userId}`, {
+    const isAdmin = await fetch(`http://localhost:8000/users/${userId}`, {
         headers: {
             "Content-Type": "application/json",
             "Accept": "application/json"
