+++ src/views/ApplicationViews.js
@@ -1,16 +1,16 @@
-import { Outlet, Route, Routes } from "react-router-dom"
-import { Login } from "../components/auth/Login"
-import { Register } from "../components/auth/Register"
-import { Authorized } from "./Authorized"
-import { PostsList } from "./PostsList"
-import { MyPosts } from "../components/posts/MyPosts.jsx"
-import { PostForm } from "../components/posts/PostForm.jsx"
-import { ConditionalPostView } from "../components/posts/ConditionalPostView.jsx"
-import { PostDetail } from "./PostDetail"
-import { UserList } from "../managers/UserList"
-import { Admin } from "./Admin.js"
-import { UnapprovedPosts } from "../components/admin/UnapprovedPosts.jsx"
-import { CategoryList } from "../components/categories/CategoryList.jsx"
+import { Outlet, Route, Routes } from "react-router-dom";
+import { Login } from "../components/auth/Login";
+import { Register } from "../components/auth/Register";
+import { Authorized } from "./Authorized";
+import { PostsList } from "../components/posts/PostsList.js";
+import { MyPosts } from "../components/posts/MyPosts.jsx";
+import { PostForm } from "../components/posts/PostForm.jsx";
+import { ConditionalPostView } from "../components/posts/ConditionalPostView.jsx";
+import { PostDetail } from "../components/posts/PostDetail.js";
+import { UserList } from "../components/admin/UserList.js";
+import { Admin } from "./Admin.js";
+import { UnapprovedPosts } from "../components/admin/UnapprovedPosts.jsx";
+import { CategoryList } from "../components/categories/CategoryList.jsx";
 
 export const ApplicationViews = ({ token, setToken }) => {
   return (
@@ -26,16 +26,18 @@ export const ApplicationViews = ({ token, setToken }) => {
           <Route path="tags/:id" element={<>Coming Soon</>} />
         </Route>
         {/* Add Routes here */}
-        <Route path="/new_post" element={<PostForm/>}/>
-        <Route path="/post/:id" element={<ConditionalPostView/>}/>
-        <Route path="/posts/:userId" element={<MyPosts/>}/>
+        <Route path="/new_post" element={<PostForm />} />
+        <Route path="/post/:id" element={<ConditionalPostView />} />
+        <Route path="/posts/:userId" element={<MyPosts />} />
         <Route path="/categories" element={<CategoryList />} />
-        <Route path="/users" element={<UserList />} />
         {/* Add Admin Routes here */}
-        <Route path="/admin" element={<Admin token={token}/>}>
-          <Route path="unapproved_posts" element={<UnapprovedPosts/>}/>
+        <Route path="/admin" element={<Admin token={token} />}>
+          <Route path="" element={<>Test</>} />
+          <Route path="unapproved_posts" element={<UnapprovedPosts />} />
+          <Route path="users" element={<UserList />} />
         </Route>
       </Route>
     </Routes>
   );
 };
+
